requirements.txt
# Lia AI Service Dependencies

# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.4

# AI Framework
openai>=1.3.0
supabase>=2.0.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
tiktoken>=0.8.0
numpy>=1.24.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx>=0.24.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
